requirements.txt
google-auth
google-cloud-storage
google-cloud-firestore
httpx
fastapi
python-multipart
pydantic-settings
uvicorn
